backend/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic
passlib[bcrypt]
python-jose
pytest
httpx
